+++ frontend/src/shared/components/form/LabeledSection.tsx
